Use float to prevent casting.
Former-commit-id: 4d0f0912927e86fb3dda26bc8131c2b61e4913d6
Former-commit-id: 87fc5322e14aa23b36a5edc91e2a539be87e6568
Former-commit-id: ba498a73a37b2550eded45e46a37515e5182d157

+++ src/ImageProcessorCore/Samplers/Resamplers/BicubicResampler.cs
@@ -13,28 +13,28 @@ namespace ImageProcessorCore
     public class BicubicResampler : IResampler
     {
         /// <inheritdoc/>
-        public double Radius => 2;
+        public float Radius => 2;
 
         /// <inheritdoc/>
-        public double GetValue(double x)
+        public float GetValue(float x)
         {
             // The coefficient.
-            double a = -0.5d;
+            float a = -0.5F;
 
-            if (x < 0)
+            if (x < 0F)
             {
                 x = -x;
             }
 
-            double result = 0;
+            float result = 0;
 
-            if (x <= 1)
+            if (x <= 1F)
             {
-                result = (((1.5d * x) - 2.5d) * x * x) + 1;
+                result = (((1.5F * x) - 2.5F) * x * x) + 1;
             }
-            else if (x < 2)
+            else if (x < 2F)
             {
-                result = (((((a * x) + 2.5d) * x) - 4) * x) + 2;
+                result = (((((a * x) + 2.5F) * x) - 4) * x) + 2;
             }
 
             return result;
