+++ src/ImageProcessorCore/Samplers/Resamplers/RobidouxSharpResampler.cs
@@ -12,13 +12,13 @@ namespace ImageProcessorCore
     public class RobidouxSharpResampler : IResampler
     {
         /// <inheritdoc/>
-        public double Radius => 2;
+        public float Radius => 2;
 
         /// <inheritdoc/>
-        public double GetValue(double x)
+        public float GetValue(float x)
         {
-            const double B = 0.2620145123990142D;
-            const double C = 0.3689927438004929D;
+            const float B = 0.2620145123990142F;
+            const float C = 0.3689927438004929F;
 
             return ImageMaths.GetBcValue(x, B, C);
         }
